Consistent types for vector inputs/outputs in base layer (#5892)
As part of an ongoing effort to clean up type hints in layers, this PR
adds consistent typing for vector input/ouputs in the base layer class.
The choices I made were:

- In public API, be as permissible as possible and use `npt.ArrayLike`
for typing, and then `np.asarray` to coerce everything into an array as
soon as possible.
- In private API, for simplicity and performance use `np.ndarray`
everywhere.
- Use `List[int]` for `dims_displayed`, as this allows `int` to be
specified, and no arithmetic needs to be done on this variable.

This fixes lots of existing typing errors in `napari.layers.base`, but
not all of them. I thought this change was self-contained enough to go
in its own PR.

---------

Co-authored-by: Wouter-Michiel Vierdag <w-mv@hotmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>